Refactored code into bucket and file-grouped methods
